refactor(itun): split pure rail derivations out of SheetRailParts
mechRailItems/pilotRailItems/crawlerRailItems/bayStates/mechStatusPill are
not components; keeping them beside RailCta made the file mix component and
non-component exports, which breaks Fast Refresh and needed a per-function
biome suppression. They move to railStats.ts and the suppressions go away.
